feat(S-13.01): fix VP-072 violations — extract test module, clean skill path references
