+++ plugins/vsdd-factory/skills/create-prd/SKILL.md
@@ -132,16 +132,12 @@ Criticality determines review depth, test coverage requirements, and holdout sce
 <Things that need resolution>
 ```
 
-### Individual BCs (`.factory/specs/behavioral-contracts/ss-NN/BC-S.SS.NNN.md`)
+### Behavioral Contracts
 
-One file per contract, sharded into per-subsystem `ss-NN/` directories,
-following `spec-format.md` format. Shard directory name is the bare
-`SS-NN` identifier (lowercased) — descriptive subsystem names live in
-ARCH-INDEX Subsystem Registry, not in the directory name.
-
-### BC Index (`.factory/specs/behavioral-contracts/BC-INDEX.md`)
-
-Table of all contracts with status tracking.
+Files are stored in `.factory/specs/behavioral-contracts/` — one file per contract
+sharded into per-subsystem `ss-NN/` directories (e.g. `ss-NN/BC-S.SS.NNN.md`).
+The BC-INDEX.md in that directory lists all contracts with status tracking.
+Shard directory name is the bare `SS-NN` identifier (lowercased).
 
 ## Self-Review (before adversarial review)
